pyproject.toml
[project]
name = "mail-reply-smith"
version = "0.1.0"
description = "Python-based mail relay and filter supporting IMAP/SMTP with Docker"
authors = [{ name="Cody", email="mathcoder23@126.com"}]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "*"
loguru = "*"